Adding BLAS_LIBS variable to R Makefile
When trying to build with openblas_openmp as the BLAS backend, it was still trying to link with -lopenblas while building, even though the pkgsrc variable BLAS_LIBS was set to -lopenblas_openmp. This was breaking the build, but adding that variable directly to the CONFIGURE_ARGS definition fixed it.